Add library module and pattern tests
